requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.13
django-cors-headers==3.11.0
django-crontab==0.7.1
et-xmlfile==1.1.0
fpdf==1.7.2
google-api-core==2.7.2
google-auth==2.6.5
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
html2text==2020.1.16
idna==3.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
natsort==8.1.0
numpy==1.21.6
opencv-python==4.5.5
openpyxl==3.0.9
pandas==1.3.5
Pillow==9.1.1
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyngrok==5.1.0
pyodbc==4.0.32
PyPDF2==1.27.12
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
reportlab==3.6.9
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
xlrd==2.0.1